[backend] separate USB from Hamilton backend (#91)
* [backend] separate USB from Hamilton backend

* refactor machine.py and USBBackend.py into top level directory

---------
