+++ pylabrobot/liquid_handling/backends/__init__.py
@@ -3,7 +3,6 @@ from .chatterbox_backend import ChatterBoxBackend
 from .serializing_backend import SerializingBackend, SerializingSavingBackend # many rely on this
 from .websocket import WebSocketBackend
 
-from .USBBackend import USBBackend
 from .hamilton.STAR import STAR
 from .hamilton.vantage import Vantage
 from .http import HTTPBackend
