+++ pylabrobot/liquid_handling/backends/backend.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from abc import ABCMeta, abstractmethod
 from typing import List, Type, Optional
 
-from pylabrobot.machine import MachineBackend
+from pylabrobot.machines.backends import MachineBackend
 from pylabrobot.resources import Resource
 from pylabrobot.liquid_handling.standard import (
   Pickup,
