+++ pylabrobot/machines/backends/__init__.py
@@ -1,0 +1,2 @@
+from .machine import MachineBackend
+from .usb import USBBackend
