+++ pylabrobot/machines/machine.py
@@ -1,10 +1,12 @@
 from __future__ import annotations
 
-from abc import ABC, ABCMeta, abstractmethod
-import functools
-from typing import Callable, Optional
+from abc import ABCMeta, abstractmethod
+from typing import Optional, Callable
 
+from pylabrobot.machines.backends import MachineBackend
 from pylabrobot.resources import Resource
+
+import functools
 
 
 def need_setup_finished(func: Callable):
@@ -22,19 +24,6 @@ def need_setup_finished(func: Callable):
       raise RuntimeError("The setup has not finished. See `setup`.")
     return await func(self, *args, **kwargs)
   return wrapper
-
-
-class MachineBackend(ABC):
-  """ Abstract class for machine backends. """
-
-  @abstractmethod
-  async def setup(self):
-    pass
-
-  @abstractmethod
-  async def stop(self):
-    pass
-
 
 class Machine(Resource, metaclass=ABCMeta):
   """ Abstract class for machine frontends. All Machines are Resources. """
