+++ pylabrobot/pumps/backend.py
@@ -1,7 +1,7 @@
 from abc import ABCMeta, abstractmethod
 from typing import List
 
-from pylabrobot.machine import MachineBackend
+from pylabrobot.machines.backends import MachineBackend
 
 
 class PumpBackend(MachineBackend, metaclass=ABCMeta):
