+++ pylabrobot/pumps/pump.py
@@ -1,7 +1,7 @@
 import asyncio
 from typing import Optional, Union
 
-from pylabrobot.machine import Machine
+from pylabrobot.machines.machine import Machine
 from .backend import PumpBackend
 from .calibration import PumpCalibration
