+++ pylabrobot/scales/scale.py
@@ -1,6 +1,6 @@
 from typing import Optional
 
-from pylabrobot.machine import Machine
+from pylabrobot.machines.machine import Machine
 from pylabrobot.scales.scale_backend import ScaleBackend
