+++ pylabrobot/temperature_controlling/temperature_controller.py
@@ -2,7 +2,7 @@ import asyncio
 import time
 from typing import Optional
 
-from pylabrobot.machine import Machine
+from pylabrobot.machines.machine import Machine
 
 from .backend import TemperatureControllerBackend
